+++ src/python_entry..rs
@@ -5,16 +5,10 @@ use std::f64::consts::PI;
 use std::sync::{Arc, Mutex};
 use std::thread;
 
-/// 2D Rastrigin function
-/// TODO: need to adjust
-#[inline]
-fn rastrigin_2d(x: f64, y: f64) -> f64 {
-    let scale = 1.5;
-    let x_scaled = x / scale;
-    let y_scaled = y / scale;
-    20.0 + x_scaled.powi(2) - 10.0 * (2.0 * PI * x_scaled).cos()
-        + y_scaled.powi(2) - 10.0 * (2.0 * PI * y_scaled).cos()
-}
+mod math;                 // declare the module
+use math::rastrigin_2d;         // import the function
+
+
 
 /// Run single SA optimization
 fn run_single_sa(
